requirements.txt
Django==3.1.13
djangorestframework==3.11.0
django-filter==21.1